:bug: Sample画像クリック時に正しいのが反映されない問題を修正(SSRで画像リストを生成するように修正)
